float and double are not reserved words
There is no floating point anywhere now: no float or double type, no
float constants, no float rules, no float helpers.  The two words came
out of the keyword table rather than staying as types that gripe, so a
program that has to port floating point code can name its own
representation with them:

	struct myflt { unsigned long bits; };
	typedef struct myflt double;

What such a program gives up is binary operators on the type - and
structure assignment and structure return, which the compiler does not
have either - so its arithmetic has to be calls taking pointers.

The assembler lost .deff with the rest.  It had been assembling float
literals to IEEE 754, which was never the right thing to build in: the
Z80 machines that shipped with a floating point unit did not agree on
a format, and neither did the BASIC ROMs.  Two of libc's float
assembly files used the directive and can no longer assemble, which is
the same answer from a different direction.

Token codes 23, 130 and 131 are left vacant rather than reused.
token_props is indexed by token number and .x streams carry these
codes, so renumbering would silently reinterpret both.

The packed keyword recognizer needed no hand editing - mkkw recomputes
the trie and its skip counts, and 'f' now holds only for, 'd' only
default and do.

Two bugs this turned up:

pass1 dereferenced a name's type without checking it was there.  A
malformed declaration leaves an entry behind with no type, and every
file that declared atof through stdlib.h then crashed c0 rather than
griping.  That was the whole of stage1 failing, not float.

libc's float sources were never rebuilt, so the tree had been resting
on stale .o files.  A clean build showed it.

libc keeps the float sources but does not build them; they are a
starting point for a rewrite against a program-supplied
representation.  fnum.c and atof.c stay in as stubs, so printf and
scanf still link and a program can override either by linking its own
- its object satisfies the reference before the linker reaches the
archive.  doprnt hands %e %f %g to _fnum by address, doscan likewise.

Sizes: cpp 44146, c0 47782, c1 47674.

+++ tests/kr_funcs.c
@@ -17,12 +17,12 @@ int noparams()
 
 // Modern ANSI style for comparison (these work)
 int modern1(int x);
-int modern2(int a, char b, float c);
+int modern2(int a, char b, long c);
 char *modern3(char *s, int n);
 
 int multi(a, b, c)
 int a;
 char b;
-float c;
+long c;
 {
 }
